@@ -55,9 +55,9 @@
     "moment": "^2.27.0",
     "prop-types": "^15.7.2",
     "query-string": "^6.13.1",
-    "react": "16.11.0",
-    "react-dom": "16.11.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.0.tar.gz",
     "react-native-animatable": "^1.3.3",
     "react-native-appearance": "~0.3.3",
     "react-native-awesome-alerts": "^1.3.4",
